Initial refactor of entrance
